refactor: consolidate render utilities
